update the readme section about creating an appscript trigger in google docs from an 'onchange' event to a time based event.
Although the on update trigger was good for adding rows immediately when needed, on days where a premium is paid into the pension pot, and 2 rows need to be added, the first empty row is added before the second 'data' row is added, creating a gap in the spreadsheet. A time based trigger, set for a time of day after the script is scheduled, adds all required rows at once, without creating gaps.